Replace the old text ingestor with a new one.
